feat: 前后端完整提交 - 科技感UI重构 + 代码注释 + 管理后台优化
主要变更:
- 前后端页面增加赛博朋克科技感主题样式
- 管理后台新增返回前台按钮,用户前台根据角色显示管理后台入口
- 后端134个Java文件全部完成逐行中文注释
- 管理后台前后台跳转动态URL适配开发和生产环境
- 更新前端README文档为项目相关内容
- 新增管理后台前端(bookstore-admin-frontend)完整项目
- 新增sql初始化脚本,Docker部署配置,开发启动脚本

+++ bookstore-common/src/main/java/com/bookstore/common/entity/UserCoupon.java
@@ -19,7 +19,7 @@ public class UserCoupon {  // 用户优惠券实体类
     @TableField("user_id")  // 字段映射注解，指定Java字段对应的数据库列名为"user_id"
     private String userId;  // 用户ID，关联account表，标识领取优惠券的用户
     @TableField("coupon_id")  // 字段映射注解，指定Java字段对应的数据库列名为"coupon_id"
-    private Integer couponId;  // 优惠券ID，关联coupon表
+    private Long couponId;  // 优惠券ID，关联coupon表
     @TableField("is_used")  // 字段映射注解，指定Java字段对应的数据库列名为"is_used"
     private Integer isUsed;  // 是否已使用：0-未使用 1-已使用
     @TableField("grant_time")  // 字段映射注解，指定Java字段对应的数据库列名为"grant_time"
